@@ -1,6 +1,6 @@
 {
   "name": "@fastify/busboy",
-  "version": "1.2.1",
+  "version": "3.1.1",
   "private": false,
   "author": "Brian White <mscdex@mscdex.net>",
   "contributors": [
@@ -17,40 +17,29 @@
   ],
   "description": "A streaming parser for HTML form data for node.js",
   "main": "lib/main",
+  "type": "commonjs",
   "types": "lib/main.d.ts",
   "scripts": {
     "bench:busboy": "cd benchmarks && npm install && npm run benchmark-fastify",
     "bench:dicer": "node bench/dicer/dicer-bench-multipart-parser.js",
     "coveralls": "nyc report --reporter=lcov",
     "lint": "npm run lint:standard",
-    "lint:everything": "npm run lint && npm run test:types",
     "lint:fix": "standard --fix",
     "lint:standard": "standard --verbose | snazzy",
-    "test:mocha": "mocha test",
+    "test:unit": "c8 --statements 98 --branches 97 --functions 96 --lines 98 node --test",
     "test:types": "tsd",
-    "test:coverage": "nyc npm run test",
-    "test": "npm run test:mocha"
-  },
-  "engines": {
-    "node": ">=14"
-  },
-  "dependencies": {
-    "text-decoding": "^1.0.0"
+    "test": "npm run test:unit && npm run test:types"
   },
   "devDependencies": {
-    "@types/node": "^18.0.0",
-    "busboy": "^1.0.0",
-    "chai": "^4.3.6",
-    "eslint": "^8.23.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-n": "^15.2.5",
-    "mocha": "^10.0.0",
-    "nyc": "^15.1.0",
+    "@types/node": "^22.0.0",
+    "busboy": "^1.6.0",
+    "c8": "^10.1.2",
     "photofinish": "^1.8.0",
     "snazzy": "^9.0.0",
-    "standard": "^17.0.0",
-    "tsd": "^0.25.0",
-    "typescript": "^4.5.2"
+    "standard": "^17.1.0",
+    "tinybench": "^3.0.0",
+    "tsd": "^0.31.0",
+    "typescript": "~5.7.2"
   },
   "keywords": [
     "uploads",
@@ -61,8 +50,12 @@
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/fastify/busboy.git"
+    "url": "git+https://github.com/fastify/busboy.git"
   },
+  "bugs": {
+    "url": "https://github.com/fastify/busboy/issues"
+  },
+  "homepage": "https://github.com/fastify/busboy#readme",
   "tsd": {
     "directory": "test/types",
     "compilerOptions": {
