@@ -54,26 +55,30 @@
     "wrapper.mjs"
   ],
   "devDependencies": {
-    "@babel/cli": "7.18.10",
-    "@babel/core": "7.18.10",
-    "@babel/eslint-parser": "7.18.9",
-    "@babel/preset-env": "7.18.10",
-    "@commitlint/cli": "17.0.3",
-    "@commitlint/config-conventional": "17.0.3",
+    "@babel/cli": "7.24.6",
+    "@babel/core": "7.24.6",
+    "@babel/eslint-parser": "7.24.6",
+    "@babel/plugin-syntax-import-attributes": "7.24.6",
+    "@babel/preset-env": "7.24.6",
+    "@commitlint/cli": "19.3.0",
+    "@commitlint/config-conventional": "19.2.2",
+    "@wdio/browserstack-service": "7.16.10",
+    "@wdio/cli": "7.16.10",
+    "@wdio/jasmine-framework": "7.16.6",
+    "@wdio/local-runner": "7.16.10",
+    "@wdio/spec-reporter": "7.16.9",
+    "@wdio/static-server-service": "7.16.6",
     "bundlewatch": "0.3.3",
-    "eslint": "8.21.0",
-    "eslint-config-prettier": "8.5.0",
-    "eslint-config-standard": "17.0.0",
-    "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-promise": "6.0.0",
-    "husky": "8.0.1",
-    "jest": "28.1.3",
-    "lint-staged": "13.0.3",
+    "eslint": "9.4.0",
+    "eslint-plugin-prettier": "5.1.3",
+    "globals": "15.3.0",
+    "husky": "9.0.11",
+    "jest": "29.7.0",
+    "lint-staged": "15.2.5",
+    "neostandard": "0.5.1",
     "npm-run-all": "4.1.5",
     "optional-dev-dependency": "2.0.1",
-    "prettier": "2.7.1",
+    "prettier": "3.3.0",
     "random-seed": "0.3.0",
     "runmd": "1.3.9",
     "standard-version": "9.5.0"
