@@ -1,26 +1,30 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import Home from './pages/Home/Home';
 import Login from './pages/Login/Login';
 import Player from './pages/Player/Player';
 import { Routes, Route, useNavigate, Navigate } from "react-router-dom";
 import { ToastContainer } from 'react-toastify';
-import { checkAuthState } from './front-firebase'; // Nouvelle fonction pour vérifier l'état de l'authentification
+import { checkAuthState } from './front-firebase';
 
 const App = () => {
   const navigate = useNavigate();
+  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    // Écoute les changements d'état de connexion de l'utilisateur
+    // Vérification de l'état d'authentification
     checkAuthState((user) => {
       if (user) {
-        console.log('Connecté');
         navigate('/home');
       } else {
-        console.log('Déconnecté');
         navigate('/login');
       }
+      setLoading(false);
     });
-  }, []);
+  }, [navigate]);
+
+  if (loading) {
+    return <div>Chargement...</div>;
+  }
 
   return (
     <div>
@@ -29,7 +33,7 @@ const App = () => {
         <Route path="/home" element={<Home />} />
         <Route path="/login" element={<Login />} />
         <Route path="/player/:id" element={<Player />} />
-        <Route path="*" element={<Navigate to="/login" />} /> {/* Redirection pour les erreurs 404 */}
+        <Route path="*" element={<Navigate to="/login" />} />
       </Routes>
     </div>
   );
